Add cat test.
