+++ python-package/xgboost/testing/data_iter.py
@@ -1,9 +1,13 @@
 """Tests related to the `DataIter` interface."""
 
+from typing import Callable, Optional
+
 import numpy as np
 
 import xgboost
 from xgboost import testing as tm
+
+from ..compat import concat
 
 
 def run_mixed_sparsity(device: str) -> None:
@@ -33,3 +37,71 @@ def run_mixed_sparsity(device: str) -> None:
     Xy_1 = xgboost.QuantileDMatrix(X_arr, y_arr)
 
     assert tm.predictor_equal(Xy_0, Xy_1)
+
+
+class CatIter(xgboost.DataIter):  # pylint: disable=too-many-instance-attributes
+    """An iterator for testing categorical features."""
+
+    def __init__(  # pylint: disable=too-many-arguments
+        self,
+        n_samples_per_batch: int,
+        n_features: int,
+        *,
+        n_batches: int,
+        n_cats: int,
+        sparsity: float,
+        onehot: bool,
+        device: str,
+        cache: Optional[str],
+    ) -> None:
+        super().__init__(cache_prefix=cache)
+        self.n_samples_per_batch = n_samples_per_batch
+        self.n_features = n_features
+        self.n_batches = n_batches
+        self.n_cats = n_cats
+        self.sparsity = sparsity
+        self.onehot = onehot
+        self.device = device
+
+        xs, ys = [], []
+        for i in range(n_batches):
+            cat, y = tm.make_categorical(
+                self.n_samples_per_batch,
+                self.n_features,
+                n_categories=self.n_cats,
+                onehot=self.onehot,
+                sparsity=self.sparsity,
+                random_state=self.n_samples_per_batch * self.n_features * i,
+            )
+            xs.append(cat)
+            ys.append(y)
+
+        self.xs = xs
+        self.ys = ys
+
+        self.x = concat(xs)
+        self.y = concat(ys)
+
+        self._it = 0
+
+    def xy(self) -> tuple:
+        """Return the concatenated data."""
+        return self.x, self.y
+
+    def next(self, input_data: Callable) -> bool:
+        if self._it == self.n_batches:
+            # return False to let XGBoost know this is the end of iteration
+            return False
+        X, y = self.xs[self._it], self.ys[self._it]
+        if self.device == "cuda":
+            import cudf  # pylint: disable=import-error
+            import cupy  # pylint: disable=import-error
+
+            X = cudf.DataFrame(X)
+            y = cupy.array(y)
+        input_data(data=X, label=y)
+        self._it += 1
+        return True
+
+    def reset(self) -> None:
+        self._it = 0
