+++ tests/python-gpu/test_gpu_data_iterator.py
@@ -92,6 +92,35 @@ def test_extmem_qdm(
         n_bins=n_bins,
         device="cuda",
         on_host=on_host,
+        is_cat=False,
+        onehot=False,
+    )
+
+
+@given(
+    strategies.integers(1, 2048),
+    strategies.integers(1, 4),
+    strategies.integers(2, 16),
+    strategies.booleans(),
+    strategies.booleans(),
+)
+@settings(deadline=None, max_examples=10, print_blob=True)
+def test_categorical_extmem_qdm(
+    n_samples_per_batch: int,
+    n_batches: int,
+    n_bins: int,
+    on_host: bool,
+    onehot: bool,
+) -> None:
+    check_extmem_qdm(
+        n_samples_per_batch,
+        4,
+        n_batches=n_batches,
+        n_bins=n_bins,
+        device="cuda",
+        on_host=on_host,
+        onehot=onehot,
+        is_cat=True,
     )
