+++ tests/python/test_data_iterator.py
@@ -328,6 +328,30 @@ def test_extmem_qdm(
         n_bins=n_bins,
         device="cpu",
         on_host=False,
+        onehot=False,
+        is_cat=False,
+    )
+
+
+@given(
+    strategies.integers(1, 4096),
+    strategies.integers(1, 4),
+    strategies.integers(2, 16),
+    strategies.booleans(),
+)
+@settings(deadline=None, max_examples=10, print_blob=True)
+def test_categorical_extmem_qdm(
+    n_samples_per_batch: int, n_batches: int, n_bins: int, onehot: bool
+) -> None:
+    check_extmem_qdm(
+        n_samples_per_batch,
+        4,
+        n_batches=n_batches,
+        n_bins=n_bins,
+        device="cpu",
+        on_host=False,
+        onehot=onehot,
+        is_cat=True,
     )
